Created auth controller for accessing db with auth routes / Created auth middleware for assigning and confirming jwt for protected routes
