+++ middleware/auth.js
@@ -1,0 +1,26 @@
+require('dotenv').config();
+const jwt = require('jsonwebtoken');
+
+const JWT_SECRET = process.env.JWT_SECRET;
+
+module.exports = function (req, res, next) {
+  // Get token from header
+  const token = req.header('x-auth-token');
+
+  // Check if no token exists
+  if (!token) {
+    return res.status(401).json({ msg: 'No token, authorization denied' });
+  }
+
+  // Verify token
+  try {
+    // Stores payload with user and token
+    const decoded = jwt.verify(token, JWT_SECRET);
+
+    // Assigns req.user to payload
+    req.user = decoded.user;
+    next();
+  } catch (err) {
+    res.status(401).json({ msg: 'Token is not valid' });
+  }
+};
